% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{get_plot}
\alias{get_plot}
\title{Retrieve a Plot by Name}
\usage{
get_plot(plot_name, figures_dir = "output/figure_objs", verbose = TRUE)
}
\arguments{
\item{plot_name}{A character string specifying the name of the plot.}

\item{figures_dir}{A character string specifying the directory to save the
plot to. The default is "output/figure_objs".}

\item{verbose}{A logical value indicating whether to print messages. The
default is TRUE.}
}
\value{
The plot object corresponding to the given plot name.
}
\description{
This function retrieves a plot object based on the provided plot name.
}
\examples{
\dontrun{
  plot <- get_plot("my_plot")
  print(plot)
}
}
