% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_handling.R
\name{get_plot_file_time}
\alias{get_plot_file_time}
\title{Get Plot File Time}
\usage{
get_plot_file_time(plot_qs_file, as_string = FALSE)
}
\arguments{
\item{plot_qs_file}{A string representing the path to the plot file.}

\item{as_string}{A logical value indicating whether to return the timestamp
as a string (TRUE) or as a POSIXct object (FALSE). Default is FALSE.}
}
\value{
A POSIXct object representing the timestamp of the plot file.
}
\description{
This function retrieves the timestamp of a given plot file.
}
\keyword{internal}
