% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_handling.R
\name{get_plot_metadata}
\alias{get_plot_metadata}
\title{Get Plot Metadata}
\usage{
get_plot_metadata(figures_dir = "output/figure_objs")
}
\arguments{
\item{figures_dir}{A character string specifying the directory to save the
plot to. The default is "output/figure_objs".}
}
\value{
A data frame containing the plot metadata if the file exists.
}
\description{
This function reads a plot metadata file and returns its contents as a data
frame.
}
\details{
The function checks if the specified plot metadata file exists. If
it does, it reads the file using \code{read.table} with tab-separated values
and returns the resulting data frame. If the file does not exist, the
function stops execution and returns an error message.
}
\keyword{internal}
