% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_handling.R
\name{get_plot_name_files}
\alias{get_plot_name_files}
\title{Get Plot Name Files}
\usage{
get_plot_name_files(plot_name, figures_dir = "output/figure_objs")
}
\arguments{
\item{plot_name}{A character string specifying the name of the plot.}

\item{figures_dir}{A character string specifying the directory to save the
plot to. The default is "output/figure_objs".}
}
\value{
A character vector containing the file names associated with the
specified plot name.
}
\description{
This function retrieves the file names associated with a given plot name.
}
\keyword{internal}
