% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_handling.R
\name{get_plot_name_from_qs_file}
\alias{get_plot_name_from_qs_file}
\title{Extract Plot Name from QS File}
\usage{
get_plot_name_from_qs_file(plot_qs_file)
}
\arguments{
\item{plot_qs_file}{A string representing the path to the QS file.}
}
\value{
A string containing the name of the plot extracted from the QS file.
}
\description{
This function retrieves the plot name from a given QS file.
}
\keyword{internal}
