% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\name{import_all_qs_plots}
\alias{import_all_qs_plots}
\title{Import All QS Plots}
\usage{
import_all_qs_plots(
  from_figures_dir,
  to_figures_dir = "output/figure_objs",
  user = "imported from file",
  cleanup = TRUE,
  overwrite_newer = FALSE
)
}
\arguments{
\item{from_figures_dir}{A character string specifying the directory
containing the qs plot files to be imported. These must have been created
by the stash_plot function.}

\item{to_figures_dir}{A character string specifying the directory to save the
imported plots to. The default is "output/figure_objs".}

\item{user}{A string representing the user who imported the plot. The default
is "imported from file".}

\item{cleanup}{A logical value indicating whether to remove old plot files
when writing out a new plot with the same plot_name. The default is TRUE.}

\item{overwrite_newer}{A logical value indicating whether to overwrite a plot
file even if the new one is older. The default is FALSE.}
}
\value{
A list of imported QS plots.
}
\description{
This function imports all QS plots from a specified directory.
}
\examples{
figures_dir <- "/path/to/figures"
plots <- import_all_qs_plots(figures_dir)

}
