% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\name{import_plot}
\alias{import_plot}
\title{Import Plot from QS File}
\usage{
import_plot(
  qs_file,
  figures_dir = "output/figure_objs",
  user = "imported from file",
  cleanup = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{qs_file}{A string representing the path to the QS file containing the
plot data.}

\item{figures_dir}{A character string specifying the directory to save the
plot to. The default is "output/figure_objs".}

\item{user}{A string representing the user who imported the plot. The default
is "imported from file".}

\item{cleanup}{A logical value indicating whether to remove old plot files
when writing out a new plot with the same plot_name. The default is TRUE.}

\item{verbose}{A logical value indicating whether to print messages. The
default is TRUE.}
}
\value{
The imported plot object.
}
\description{
This function imports a plot from a specified QS file.
}
\examples{
# Example usage:
plot <- import_plot("path/to/your/plot.qs")

}
