% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{list_plots}
\alias{list_plots}
\title{List Plot Files}
\usage{
list_plots(figures_dir = "output/figure_objs")
}
\arguments{
\item{figures_dir}{A character string specifying the directory where plot
files are stored. Default is "output/figure_objs".}
}
\value{
A character vector of file names in the specified directory.
}
\description{
This function lists all plot files in the specified directory.
}
\examples{
\dontrun{
  list_plots()
  list_plots("custom/figure_objs")
}

}
