% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_handling.R
\name{metadata_as_df}
\alias{metadata_as_df}
\title{Convert Metadata List to Data Frame}
\usage{
metadata_as_df(metadata)
}
\arguments{
\item{metadata}{A list of metadata from get_plot_metadata()}
}
\value{
A data frame containing the metadata.
}
\description{
This function takes metadata and converts it into a data frame.
}
\keyword{internal}
