% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_handling.R
\name{plot_file_name_exists}
\alias{plot_file_name_exists}
\title{Check if a Plot File Name Exists}
\usage{
plot_file_name_exists(plot_name, figures_dir = "output/figure_objs")
}
\arguments{
\item{plot_name}{A character string specifying the name of the plot.}

\item{figures_dir}{A character string specifying the directory to save the
plot to. The default is "output/figure_objs".}
}
\value{
A logical value indicating whether the plot file name exists (TRUE)
or not (FALSE).
}
\description{
This function checks whether a plot file with the given name already exists.
}
\keyword{internal}
