% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\name{remove_old_plot}
\alias{remove_old_plot}
\title{Remove Old Plot}
\usage{
remove_old_plot(plot_name, figures_dir = "output/figure_objs", cleanup = TRUE)
}
\arguments{
\item{plot_name}{A character string specifying the name of the plot.}

\item{figures_dir}{A character string specifying the directory to save the
plot to. The default is "output/figure_objs".}

\item{cleanup}{A logical value indicating whether to remove old plot files
when writing out a new plot with the same plot_name. The default is TRUE.}
}
\value{
None. This function is called for its side effects.
}
\description{
This function removes an old plot file if it exists.
}
\keyword{internal}
