% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_plot}
\alias{remove_plot}
\title{Remove a Plot}
\usage{
remove_plot(plot_name, figures_dir = "output/figure_objs", verbose = TRUE)
}
\arguments{
\item{plot_name}{A character string specifying the name of the plot.}

\item{figures_dir}{A character string specifying the directory to save the
plot to. The default is "output/figure_objs".}

\item{verbose}{A logical value indicating whether to print messages. The
default is TRUE.}
}
\value{
None. The function is called for its side effect of removing the plot.
}
\description{
This function removes a plot with the specified name from your stash.
}
\examples{
\dontrun{
  # Remove a plot named "my_plot" from the current environment
  remove_plot("my_plot")

  # Remove a plot named "my_plot" from a specified environment
  remove_plot("my_plot", figures_dir = "custom_dir/figures")
}

}
