% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashPlot-package.R
\docType{package}
\name{stashPlot-package}
\alias{stashPlot}
\alias{stashPlot-package}
\title{stashPlot: A Package to Stash and Manage Plot Objects with Metadata, Allowing Easy Reloading, Modification, and Assembly of Plots}
\description{
A package to stash and manage plot objects with metadata, allowing easy reloading, modification, and assembly of plots. This package is designed to help manage the complexity of creating and modifying plots in R, especially when the plots are created in multiple steps or when the plots are created by multiple people. The package provides a way to save plot objects to disk, along with metadata about the plot. The package also provides a way to reload the plot objects and metadata so one can modify the plots and assemble the plots into a final plot. The package is designed to work with the 'ggplot2' package, but it can be used with other plotting packages as well.
}
\author{
\strong{Maintainer}: Matthew Cannon \email{matthewvc1@gmail.com}

}
\keyword{internal}
